+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 See `UPGRADE.md` for instructions on upgrading between versions.
 
+## 0.45.0
+
+* Add `Brando.Datasource`. Allows you to access preset backend queries from Villain.
+* Simplify Villain default parser. Now you can `use Brando.Villain.Parser`
+for sensible defaults, and override when neccessary.
+
+
 ## 0.44.0
 
 * Drag and drop sequence pages.